requirements.txt
pip >= 24.0.0

pylint==3.0.0
pytest==7.4.2
pytest-cov==4.1.0
black==23.9.1
#ipython==8.17.2

pandas

click

colorlog

openpyxl
xlrd
tabula-py
tabulate

matplotlib

jpype1
Unidecode

PyQt6==6.8.0
PyQt6-Qt6
pyqtgraph